% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{add_bound_comparison}
\alias{add_bound_comparison}
\title{Adiciona Comparacao De Bound Ao Corpo De Funcao}
\usage{
add_bound_comparison(body, bound, kind = c("lower", "upper"))
}
\arguments{
\item{body}{corpo da funcao, como uma expressao}

\item{bound}{valor numerico que sera usado na comparacao do corpo da funcao}

\item{kind}{um de \code{c("lower", "upper")}, indicando se a comparacao sera \code{f(x) >= bound} ou
\code{f(x) <= bound}}
}
\value{
corpo alterado, como uma expressao
}
\description{
Altera corpo \code{body} para incluir comparacao com \code{bound} no formato \code{f(x) <= 0}
}
