% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{build_function}
\alias{build_function}
\title{Constroi Funcao A Partir De Argumentos E Corpo}
\usage{
build_function(args, body)
}
\arguments{
\item{args}{lista de argumentos para a funcao}

\item{body}{corpo da funcao, como uma expressao}
}
\value{
funcao gerada
}
\description{
Gera uma funcao com argumentos \code{args} e corpo \code{body}
}
