% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{clean_expression_string}
\alias{clean_expression_string}
\title{Limpa String De Expressao Matematica}
\usage{
clean_expression_string(x, early = FALSE)
}
\arguments{
\item{x}{string com expressao para limpar}

\item{early}{logico indicando se deve parar antes de remover operadores de comparacao}
}
\value{
string limpa
}
\description{
Remove parenteses externos e espacos, e opcionalmente remove operadores e bounds de comparacao
}
