% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulascheias-package.r
\docType{package}
\name{copulascheias-package}
\alias{copulascheias}
\alias{copulascheias-package}
\title{copulascheias: Ferramental Para Analise De Parametros De Cheias Atraves De Vine-Copulas}
\description{
Ferramental Para Analise De Parametros De Cheias Atraves De Vine-Copulas.
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lucaskhenayfis@gmail.com}

}
\keyword{internal}
