% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{event2bounds}
\alias{event2bounds}
\alias{event2bounds.unitary_event_u}
\alias{event2bounds.unitary_event_m}
\title{Auxiliar Para Conversao De Eventos Em Bounds}
\usage{
event2bounds(x, ...)

\method{event2bounds}{unitary_event_u}(x, mode = c("x", "u"), ...)

\method{event2bounds}{unitary_event_m}(x)
}
\arguments{
\item{x}{evento unitario}
}
\description{
Auxiliar Para Conversao De Eventos Em Bounds
}
