% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{event2function}
\alias{event2function}
\alias{event2function.unitary_event_u}
\alias{event2function.unitary_event_m}
\title{Extrai Funcao De Evento Unitario}
\usage{
event2function(x, ...)

\method{event2function}{unitary_event_u}(x, ...)

\method{event2function}{unitary_event_m}(x, ...)
}
\arguments{
\item{x}{evento unitario}

\item{...}{demais parametros de cada metodo}
}
\value{
lista com funcoes geradas ou erro se univariado
}
\description{
Para eventos unitaros univariados, retorna erro. Para multivariados, retorna lista com funcoes
que representam a avaliacao no formato \code{f(x) <= 0} do evento descrito. Por exemplo, para o evento
"var1 + var2^2 >= 100", a funcao retornada sera \code{function(var1, var2) 100 - (var1 + var2^2)}. Se
haviam bounds inferiores e superiores, duas funcoes serao retornadas, uma para cada bound.
}
