% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{extract_variables_from_expr}
\alias{extract_variables_from_expr}
\title{Extrai Variaveis De Expressao AST}
\usage{
extract_variables_from_expr(expr)
}
\arguments{
\item{expr}{expressao parseada como objeto language}
}
\value{
vetor de caracteres com nomes das variaveis encontradas
}
\description{
Recursivamente extrai nomes de simbolos de uma expressao parseada (AST)
}
