% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{fillvec}
\alias{fillvec}
\title{Preenche Vetor Com Valores Padrao}
\usage{
fillvec(vec, names, fill = 1)
}
\arguments{
\item{vec}{vetor nomeado a ser preenchido}

\item{names}{vetor de nomes que \code{vec} deve ter}

\item{fill}{valor com o qual preencher posicoes faltantes}
}
\value{
vetor \code{vec} preenchido com \code{fill} para todas as posicoes em \code{names} faltantes em \code{vec}
}
\description{
Preenche \code{vec} com valores \code{fill} para todas as posicoes em \code{names} que nao existam em \code{vec}
}
