% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{get_bounds}
\alias{get_bounds}
\alias{get_lower}
\alias{get_upper}
\title{Extrai Bounds De Expressao}
\usage{
get_lower(x, modelo)

get_upper(x, modelo)
}
\arguments{
\item{x}{string com expressao de evento unitario}

\item{modelo}{modelo com referencia as variaveis disponiveis}
}
\value{
vetor numerico de dois elementos: lower e upper bounds, NA se inexistente na expressao
}
\description{
Extrai bounds inferior e superior de expressao do tipo "var <= b", "var >= b" ou "a <= var <= b".
}
