% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{inference2bounds}
\alias{inference2bounds}
\alias{inference2bounds.simple_inference}
\alias{inference2bounds.complex_inference}
\alias{inference2bounds.void_inference}
\title{Extrai Limites De Integracao Retangulares}
\usage{
inference2bounds(inference, modelo, ...)

\method{inference2bounds}{simple_inference}(inference, modelo, mode = "u", ...)

\method{inference2bounds}{complex_inference}(inference, modelo, ...)

\method{inference2bounds}{void_inference}(inference, modelo, mode = "u", ...)
}
\arguments{
\item{inference}{um objeto do tipo \code{simple_inference}}

\item{modelo}{modelo ajustado por \code{\link[=fit_modelo_cheia]{fit_modelo_cheia()}} para identificacao das variaveis que nao
aparecem em \code{inference}}

\item{...}{demais parametros de cada metodo}

\item{mode}{um de \code{c("x", "u")}, indicando a escala em que bounds serao retornados}
}
\value{
lista de dois vetores: lower e upper bounds de integracao
}
\description{
Gera lista de dois vetores: lower e upper bounds de integracao
}
\details{
O comportamento desta funcao depende do tipo de \code{inference}. No caso de \code{simple_inference}, a
funcao simplesmente retorna os bounds de cada evento unitario, preenchendo com 0 e 1 para
as variaveis que nao aparecem em \code{inference}.

No caso de \code{complex_inference}, a funcao identifica a bounding box que contem o evento complexo e
retorna os limites desta caixa.

E recomendado sempre passar este argumento, do contrario erros podem decorrer.
}
