% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{inference_constructors}
\alias{inference_constructors}
\alias{new_simple_inference}
\alias{new_complex_inference}
\title{Construtor Interno De Objetos De Inferencia}
\usage{
new_simple_inference(unitary_events)

new_complex_inference(unitary_events)
}
\arguments{
\item{unitary_events}{lista de eventos unitarios ja interpretados por \code{\link[=parse_unitary_event]{parse_unitary_event()}}}
}
\value{
argumento \code{unitary_events} com classe adicionada \code{simple_inference}
}
\description{
Gera um objeto para inferencia, uma lista de eventos unitarios interpretados
}
