% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{minicheias}
\alias{minicheias}
\title{Exemplo De Dado Para Modelagem}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 83 rows and 6 columns.
}
\usage{
minicheias
}
\description{
Dado de cheias multivariadas reduzido para exemplos no pacote
}
\details{
Contem dois postos e tres aspectos de cheia para cada, organizadas nas colunas nomeadas
\code{"posto_caracteristica"}
}
\keyword{datasets}
