% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{names2args}
\alias{names2args}
\title{Gera Lista De Argumentos De Funcao}
\usage{
names2args(argnames)
}
\arguments{
\item{argnames}{vetor de nomes de argumentos}
}
\value{
lista de argumentos com nomes \code{argnames} e valores \code{NULL}
}
\description{
Gera lista de argumentos com nomes \code{argnames} e valores \code{NULL} para composicao de funcoes
}
