% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{new_complex_inference}
\alias{new_complex_inference}
\title{Construtor Interno De \code{complex_inference}}
\usage{
new_complex_inference(unitary_events, separators)
}
\description{
Ainda nao implementado
}
