% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{new_simple_inference}
\alias{new_simple_inference}
\title{Construtor Interno De \code{simple_inference}}
\usage{
new_simple_inference(unitary_events)
}
\arguments{
\item{unitary_events}{lista de eventos unitarios ja interpretados por \code{\link[=parse_unitary_event]{parse_unitary_event()}}}
}
\value{
argumento \code{unitary_events} com classe adicionada \code{simple_inference}
}
\description{
Gera um objeto da classe \code{simple_inference}, uma lista de eventos unitarios interpretados
}
