% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{new_unitary_event}
\alias{new_unitary_event}
\alias{new_unitary_event_u}
\alias{new_unitary_event_m}
\title{Constroi Objeto Evento Unitario}
\usage{
new_unitary_event_u(lower, upper, var, modelo)

new_unitary_event_m(lower, upper, vars, expr, modelo)
}
\arguments{
\item{lower}{bound inferior do evento}

\item{upper}{bound superior do evento}

\item{var}{nome da variavel do evento (univariado)}

\item{modelo}{modelo com referencia as variaveis disponiveis}

\item{vars}{vetor de nomes das variaveis do evento (multivariado)}

\item{expr}{string com expressao matematica do evento (multivariado)}
}
\value{
objeto do tipo \code{unitary_event_u} ou \code{unitary_event_m}
}
\description{
Construtores internos para eventos unitarios univariados e multivariados
}
