% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{parse_inference}
\alias{parse_inference}
\title{Parse De Expressoes De Inferencia}
\usage{
parse_inference(expr, modelo)
}
\arguments{
\item{expr}{uma string indicando a inferencia desejada no modelo. Veja Exemplos e Detalhes}

\item{modelo}{o modelo no qual sera realizada a inferencia}
}
\description{
Interpreta uma expressao de inferencia e retorna objeto indicando tipo de integracao necessario
}
