% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{parse_unitary_expr}
\alias{parse_unitary_expr}
\title{Interpreta Evento Unitario}
\usage{
parse_unitary_expr(x, modelo)
}
\arguments{
\item{x}{string no formato \verb{"lhs (>=|<=) rhs} indicando um evento para inferencia}
}
\description{
Transforma string \code{x} em objeto para posterior construcao dos volumes de integracao
}
