% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{parse_validate_math_expression}
\alias{parse_validate_math_expression}
\title{Valida E Extrai Informacoes De Expressao Matematica}
\usage{
parse_validate_math_expression(x, modelo)
}
\arguments{
\item{x}{string com expressao matematica para validar}

\item{modelo}{modelo com referencia as variaveis disponiveis}
}
\value{
lista com expressao limpa e vetor de variaveis encontradas
}
\description{
Valida sintaxe da expressao matematica e extrai variaveis utilizadas
}
