% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{simple_inference2bounds}
\alias{simple_inference2bounds}
\title{Extrai Limites De Integracao De \code{simple_inference}}
\usage{
simple_inference2bounds(inference, modelo = NULL)
}
\arguments{
\item{inference}{um objeto do tipo \code{simple_inference}}

\item{modelo}{opcionalmente um modelo ajustado por \link{fit_modelo_cheia}. Veja Detalhes}
}
\value{
lista de dois vetores: lower e upper bounds de integracao
}
\description{
Gera lista de dois vetores: lower e upper bounds de integracao
}
\details{
Se \code{modelo} for passado, a funcao busca as variaveis contidas nele que nao constam em \code{inference}
para completar os vetores de bounds. Nestas posicoes o lower bound e sempre 0 e o upper, 1.

E recomendado sempre passar este argumento, do contrario erros podem decorrer.
}
