% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{split_eventos_unitarios}
\alias{split_eventos_unitarios}
\alias{split_separadores}
\title{Separadores De Trechos De Expressoes}
\usage{
split_eventos_unitarios(expr)

split_separadores(expr)
}
\arguments{
\item{expr}{uma string de evento para inferencia. Veja \code{\link[=parse_inference]{parse_inference()}}}
}
\value{
vetor de variaveis de evento unitario ou operadores de combinacao. Veja Detalhes
}
\description{
Separam uma expressao de inferencia entre expressoes de eventos unitarios e operadores de combinacao
}
\details{
Ambas as funcoes recebem uma expressao do tipo \code{"(var1 <= b1) & (var2 >= b2) ..."}.
\code{split_eventos_unitarios} retorna um vetor \code{c("var1", "var2", ...)} enquanto \code{split_separadores}
retorna os operadores \code{c("&, ...")}. O numero de operadores e sempre o numero de eventos menos 1.
}
