% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{str2function}
\alias{str2function}
\title{Gera Funcao A Partir De String}
\usage{
str2function(str, vars, bound = NA, kind = c("lower", "upper"))
}
\arguments{
\item{str}{string com expressao que sera o corpo da funcao}

\item{vars}{vetor de nomes de variaveis que podem ser usadas em \code{str}}

\item{bound}{valor numerico que sera usado na comparacao do corpo da funcao}

\item{kind}{um de \code{c("lower", "upper")}, indicando se a comparacao sera \code{f(x) >= bound} ou
\code{f(x) <= bound}}
}
\value{
funcao gerada
}
\description{
Gera uma funcao com corpo baseado em \code{str} e \code{bounds} e argumentos \code{vars}
}
\details{
\code{str2function} gera funces \code{g(x)} para um contexto de comparacao \code{g(x) <= 0}. A funcao \code{g(x)}
sera montada a partir de \code{str}, que deve ser uma expressao valida em R, e \code{bound}. Se \code{kind} for
\code{"lower"}, a funcao sera \code{g(x) = bound - (str)} e, se for \code{"upper"}, \code{g(x) = (str) - bound}.
}
