% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{validate_function}
\alias{validate_function}
\title{Valida Se \code{fun} Roda Adequadamente}
\usage{
validate_function(fun, vars, str)
}
\arguments{
\item{fun}{funcao a ser testada}

\item{vars}{vetor de nomes de variaveis que podem ser usadas em \code{str}}

\item{str}{string original que deu origem a \code{fun}, para mensagens de erro}
}
\value{
NULL invisivel se \code{fun} roda adequadamente, erro caso contrario
}
\description{
Tenta executar \code{fun} com argumentos dummy; se falhar, retorna erro
}
