% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{vars_in_str}
\alias{vars_in_str}
\title{Extrai Variaveis De String

Filtra \code{vars} para aquelas que aparecem em \code{str}}
\usage{
vars_in_str(str, vars)
}
\arguments{
\item{str}{string com expressao que sera analisada}

\item{vars}{vetor de nomes de variaveis para procurar}
}
\value{
subset dos elementos de \code{vars} que aparecem em \code{str}
}
\description{
Extrai Variaveis De String

Filtra \code{vars} para aquelas que aparecem em \code{str}
}
