% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitary_event.r
\name{event2bounds}
\alias{event2bounds}
\alias{event2bounds.unitary_event_u}
\alias{event2bounds.unitary_event_m}
\title{Extrai Bounds De Evento Unitario}
\usage{
event2bounds(x, ...)

\method{event2bounds}{unitary_event_u}(x, mode = c("x", "u"), ...)

\method{event2bounds}{unitary_event_m}(x)
}
\arguments{
\item{x}{evento unitario}

\item{...}{demais parametros de cada metodo}

\item{mode}{um de \code{c("x", "u")}, indicando a escala em que bounds serao retornados}
}
\value{
vetor numerico de dois elementos: lower e upper bounds ou erro se multivariado
}
\description{
Extrai Bounds De Evento Unitario
}
