% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo_cheia.r
\name{fit_modelo_cheia}
\alias{fit_modelo_cheia}
\title{Ajuste De Modelo De Cheias}
\usage{
fit_modelo_cheia(
  data,
  postos,
  anos = NULL,
  vine_params = default_vine_params()
)
}
\arguments{
\item{data}{\code{data.table} de dados de cheias}

\item{postos}{vetor de postos a serem modelados}

\item{anos}{vetor indicando quais anos devem ser considerados para o modelo; \code{NULL} usa todos os
anos disponiveis em \code{data}}

\item{vine_params}{lista de argumentos para estimacao da vine copula, pode ser composta de
qualquer argumento da funcao \code{\link[VineCopula:RVineStructureSelect]{VineCopula::RVineStructureSelect()}}}
}
\value{
objeto \code{modelo_cheias}
}
\description{
Estima uma vine copula considerando todos os aspectos de cheia em \code{data}
}
