% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.r
\name{new_complex_inference}
\alias{new_complex_inference}
\title{Construtor Interno De \code{complex_inference}}
\usage{
new_complex_inference(unitary_events, separators)
}
\arguments{
\item{unitary_events}{lista de eventos unitarios ja interpretados por \code{\link[=parse_unitary_expr]{parse_unitary_expr()}}}

\item{separators}{vetor de separadores de eventos}
}
\description{
Ainda nao implementado
}
