% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcopula.r
\name{pcopula}
\alias{pcopula}
\title{Inferencia Em Modelos De Copula}
\usage{
pcopula(expr, modelo)
}
\arguments{
\item{expr}{string descrevendo o evento cuja probabilidade deve ser calculada}

\item{modelo}{modelo de copula como retornado por \code{\link[=fit_modelo_cheia]{fit_modelo_cheia()}}}
}
\description{
Calcula a probabilidade de um evento \code{expr} de acordo com a copula estimada \code{modelo}
}
