% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcopula.r
\name{run_inference}
\alias{run_inference}
\alias{run_inference.simple_inference}
\title{Executa Inferencia}
\usage{
run_inference(inference, modelo, nsims)

\method{run_inference}{simple_inference}(inference, modelo, nsims = 2e+06)
}
\arguments{
\item{inference}{objeto inferenca}

\item{modelo}{modelo ajustado no qual computar inferencia}

\item{nsims}{numero de amostras de Monte Carlo}
}
\description{
Funcao interna para integracao de PDfs estimadas
}
