% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcopula.r
\name{run_inference}
\alias{run_inference}
\alias{run_inference.simple_inference}
\alias{run_inference.complex_inference}
\title{Executa Inferencia}
\usage{
run_inference(inference, modelo, nsims, modo)

\method{run_inference}{simple_inference}(inference, modelo, nsims, modo)

\method{run_inference}{complex_inference}(inference, modelo, nsims, modo)
}
\arguments{
\item{inference}{objeto inferenca}

\item{modelo}{modelo ajustado no qual computar inferencia}

\item{nsims}{numero de amostras de Monte Carlo}

\item{modo}{um de \code{c("default", "lhs")} indicando se a simulacao de Monte Carlo deve ser feita
por amostragem comum ou hipercubo latino}
}
\description{
Funcao interna para integracao de PDfs estimadas
}
