% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcopula.r
\name{run_inference}
\alias{run_inference}
\title{Executa Inferencia}
\usage{
run_inference(inference, modelo)
}
\arguments{
\item{inference}{objeto inferenca}
}
\description{
Funcao interna para integracao de PDfs estimadas
}
